added middleware for authentication

+++ resolvers/booking.js
@@ -4,7 +4,10 @@ const {transformBooking,transformEvent} =require('./merge')
 
 module.exports = {
   
-    bookings: async () => {
+    bookings: async (args,req) => {
+      if (!req.isAuth) {
+        throw new Error('Unauthenticated!');
+      }
       try {
         const bookings = await Booking.find();
         return bookings.map(booking => {
@@ -14,7 +17,10 @@ module.exports = {
         throw err;
       }
     },
-    bookEvent: async args => {
+    bookEvent: async (args,req) => {
+      if (!req.isAuth) {
+        throw new Error('Unauthenticated!');
+      }
       const fetchedEvent = await Event.findOne({ _id: args.eventId });
       const booking = new Booking({
         user: '630da3f0faf842bf711f7415',
@@ -24,8 +30,11 @@ module.exports = {
       return transformBooking(result)
     },
     cancelBooking: async args => {
+      if (!req.isAuth) {
+        throw new Error('Unauthenticated!');
+      }
       try {
-        const booking = await Booking.findById(args.bookingId).populate('event');
+        const booking = await Booking.findById(args.bookingId).populate('event');//popluate is used for het the whole object not only objectid
         const event = transformEvent(booking.event)
         await Booking.deleteOne({ _id: args.bookingId });
         return event;
@@ -33,4 +42,4 @@ module.exports = {
         throw err;
       }
     }
-  };
+  };
